Punch-vest - Arma sqf and Arduino codes
Codes run a punch-vest and/or produce ingame text info of bullet hits. Arma sqf code is in the form of init.sqf, it can also be pasted in console. The script is possibly being converted to a separate mod also. The Arduino sketch needs to be loaded in Arduino for the vest. Sqf script detects bullet hits to your body. Signal lasting 0.1 second is sent to a relay to open valve/trigger solenoid/produce electric shock or whatever you can imagine giving you punch. Depending on a caliber of the enemy weapon fired the signal is given 1 to 4 times in a row. Shooting with your own rifle produces light punch resembling recoil on your right shoulder. ARMA COM module by Fat_Lurch and GoogleBen is the engine needed behind this connection. Get it from Steam Arma workshop.